Refactor parsing structure
